# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Adversary backend API — scenarios, geo, runs (WebSocket ops+map), manage, generate/route, AUTH_MODE=off|session. Live channels: GET /v1/runs/{id}/ws/ops and /ws/map (Upgrade). Prometheus scrape: GET /metrics (not listed as OpenAPI paths). Architecture: docs/api.md.",
        "title": "Adversary API",
        "contact": {},
        "version": "0.9.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/admin/geo/jobs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-geo"
                ],
                "summary": "Reseed job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_geoseed.Job"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/admin/geo/meta": {
            "get": {
                "description": "Last seed job status, MBTiles path health, and row counts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-geo"
                ],
                "summary": "Geo catalogue metadata",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.metaResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/admin/geo/reseed": {
            "post": {
                "description": "Mines MBTILES_PATH asynchronously into PostGIS; one job at a time",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-geo"
                ],
                "summary": "Start geo catalogue reseed",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.reseedResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
...
        "/v1/geo/aerodromes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "List aerodromes in bbox",
                "parameters": [
                    {
                        "type": "number",
                        "description": "West",
                        "name": "west",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "South",
                        "name": "south",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "East",
                        "name": "east",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "North",
                        "name": "north",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max rows",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_geoseed.PointRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/geo/ports": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "List ports in bbox",
                "parameters": [
                    {
                        "type": "number",
                        "description": "West",
                        "name": "west",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "South",
                        "name": "south",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "East",
                        "name": "east",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "North",
                        "name": "north",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max rows",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_geoseed.PointRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/geo/regions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "List geo regions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_geoseed.Region"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/geo/road-anchors": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "List road anchors in bbox",
                "parameters": [
                    {
                        "type": "number",
                        "description": "West",
                        "name": "west",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "South",
                        "name": "south",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "East",
                        "name": "east",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "North",
                        "name": "north",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max rows",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_geoseed.PointRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/geo/sea-lanes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geo"
                ],
                "summary": "List sea-lane points in bbox",
                "parameters": [
                    {
                        "type": "number",
                        "description": "West",
                        "name": "west",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "South",
                        "name": "south",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "East",
                        "name": "east",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "North",
                        "name": "north",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max rows",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_geoseed.PointRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_shammalie_adversary_apps_api_internal_geoseed.Job": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mbtilesPath": {
                    "type": "string"
                },
                "progress": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_shammalie_adversary_apps_api_internal_geoseed.PointRow": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "eleFt": {
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "icao": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "regionId": {
                    "type": "string"
                },
                "runways": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "github_com_shammalie_adversary_apps_api_internal_geoseed.Region": {
            "type": "object",
            "properties": {
                "bbox": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "supports": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_handler.errorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.metaResponse": {
            "type": "object",
            "properties": {
                "aerodromeCount": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "generatedAt": {
                    "type": "string"
                },
                "jobStatus": {
                    "type": "string"
                },
                "latestJob": {
                    "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_geoseed.Job"
                },
                "mbtilesExists": {
                    "type": "boolean"
                },
                "mbtilesPath": {
                    "type": "string"
                },
                "mbtilesSourceUrl": {
                    "type": "string"
                },
                "portCount": {
                    "type": "integer"
                },
                "regionCount": {
                    "type": "integer"
                },
                "roadAnchorCount": {
                    "type": "integer"
                },
                "schemaV": {
                    "type": "integer"
                },
                "seaLaneCount": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.reseedResponse": {
            "type": "object",
            "properties": {
                "jobId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "adversary_session",
            "in": "cookie"
        }
    }
}